Use base SHA when using git diff to list modified files
Previously, the SHA of the working merge branch was comapred with the
SHA of the base branch to determine the list of files that were changed
in a particular pull request. This ended up also included any merged
changes to the base branch since the pull request was made.

Using the remote branch SHA will ensure that only the files with changes
that the remote branch includes will be returned from git diff commands.