router DC changes for setting x_forwarded_for header
With OpenShift 3.3.1.11 and 3.4.1, the OpenShift router image accepts the environment var ROUTER_USE_PROXY_PROTOCOL to have the router pod use the PROXY protocol.

This allows propogation of the X-FORWARDED-FOR HTTP header showing the client IP address instead of the AWS ELB's IP.

This will be the default setting going forward. This requires the AWS ELB to be configured to speak the PROXY protocol as well (https://docs.aws.amazon.com/elasticloadbalancing/latest/classic/enable-proxy-protocol.html).

For GCP ELBs, there is no need for this var, so only apply the extra env var when cloud is 'aws'.