Fix CI parent branch check
Previously, the list of commits found by the parent check script was
incorrect, causing failures on pull requests submitted to the prod
branch. This change reaches out to github via the v3 api with an
authenticated request to gather the definitive list of commits for a
pull request and compare them against the commits in the stg branch
before merging to prod.