Remove body and title from pull request data
The body and title of a pull request can contain any character,
including unicode characters. This poses problems when parsing json,
when passing data through as env variables to the jenkins plugin, when
passing the env variables to and openshift build, and when openshift
passes them to docker.

There are simply too many possibilities for unexpected issues to occur
when parsing these fields. They are not needed for testing and I do not
see a significant need for them in the future, so they are removed.