Fix issue with un-testable builds failing
Builds initiated from comments lacking the "[test]" string were failing
due to incorrectly testing for an empty string rather than a null value.

Also, only post a failure status once, in the Jenkinsfile. It will be
the testing script's responsibility to post the success status, to
ensure it is posted as soon as possible.