Install correct openshift-dedicated-scripts versios.
Previously we just installed the latest, so a new 3.4 cluster could now
pick up a 3.5 version of openshift-scripts-dedicated, and this would
fail due to very recent changes in the rpm.

Instead if the package is not already installed, we query the correct
major OpenShift version, and find a matching version of scripts to
install. If none matches we consider this a failure. If already
installed we skip all of this to save time.

This new role becomes a dependency for the dedicated admin, and volume
provisioner roles, both of which require this package to be installed.