Use external flock instead of python based version
When a server is nearing overload, the cron jobs might take a long time
to start up just to be able to check the flock file given the overhead
of python.  If they can't start up quick enough, then they'll just keep
piling up as the cron deamon does not wait for the previous one to
finish.

By using the external flock command, and using /host/run/lock/ for the
location of the lock file, we avoid the python overhead of startup and
place the lock file in local memory.

Note that /host/run is the host's /run tmpfs mounted into the container.