Refactored how ansible_inventory manages known_hosts
Found first that command wouldn't cut it.
Next, if you have a single shell that does it all, it's always reported as a change.
Alternative is changed_when: false, so never a changes.
To check each host, have to put in task file that is included, as
block doesn't support with_items.

So now, it will change known_hosts only when it needs to and report each update as a change.