Fix ordering of arguments passed to send_metrics
The ordering of the arguments was causing the route and service
http check items to swap values with each other. IE, a failure in
the route would cause a service error, and vice versa. This commit
fixes that issue and adds debug information to make it easier to
debug issues like this in the future.